requirements.txt
pandas
python-dateutil
pydantic
uuid
